+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 ### Fix
 - Fix trimming of CMYK images.
 - Respond with 404 when the source image can not be found in OpenStack Object Storage.
+- Respond with 404 when file wasn't found in the GCS storage.
 
 ## [3.6.0] - 2022-06-13
 ### Add